+++ src/app/(unauthenticated)/home/page.jsx
@@ -8,16 +8,14 @@ import "./page.css";
 import mockedValues from "../../../data/mockedValues.json";
 import PaginatedResults from "../../../components/paginated-results/paginated-results";
 import FiltrarApresentacaoModal from "../../../components/modal/filtrar-apresentacao/filtrarApresentacaoModal";
+import useSuap from "../login/useSuap";
 
 const Home = () => {
   const { t, i18n } = useTranslation();
-  const navigate = useNavigate();
   const [works, setWorks] = useState(mockedValues.trabalhos);
   const [isFilterModalOpen, setIsFilterModalOpen] = useState(false);
   const [isPresentationModalOpen, setIsPresentationModalOpen] = useState(false);
-  const [isProcessing, setIsProcessing] = useState(false);
-  const location = useLocation();
-  const [error, setError] = useState(null);
+  const { handleOAuthCallback } = useSuap();
   const [activeFilters, setActiveFilters] = useState({
     trabalho: {
       artigo: false,
@@ -32,158 +30,14 @@ const Home = () => {
     },
   });
 
-  const [initialHash] = useState(() => {
-    const hash = window.location.hash;
-    console.log("Initial hash captured:", hash);
-    return hash;
-  });
-
   useEffect(() => {
-    console.log("Checking for OAuth callback");
-    let hashToProcess = window.location.hash;
-    console.log("Current hash:", hashToProcess);
-
-    if (!hashToProcess) {
-      const storedHash = sessionStorage.getItem("oauth_hash");
-      console.log("Stored hash from sessionStorage:", storedHash);
-
-      if (storedHash) {
-        hashToProcess = storedHash;
-        console.log("Using stored hash:", hashToProcess);
-        sessionStorage.removeItem("oauth_hash");
-      }
+    const storedHash = sessionStorage.getItem("oauth_hash");
+    if (storedHash) {
+      hashToProcess = storedHash;
+      sessionStorage.removeItem("oauth_hash");
     }
-
-    console.log("Final hash to process:", hashToProcess);
-
-    if (hashToProcess && hashToProcess.includes("access_token=")) {
-      console.log("Processing OAuth callback with hash:", hashToProcess);
-      window.location.hash = hashToProcess;
-      handleOAuthCallback();
-    }
+    handleOAuthCallback();
   }, []);
-
-  const handleOAuthCallback = async () => {
-    const hash = window.location.hash.substring(1);
-    console.log("Full hash:", hash);
-    const params = {};
-    hash.split("&").forEach((param) => {
-      const [key, value] = param.split("=");
-      if (key && value) {
-        params[key] = decodeURIComponent(value);
-      }
-    });
-
-    console.log("Parsed params:", params);
-    console.log("access_token (underscore):", params.access_token);
-    console.log("accessToken (camelCase):", params.accessToken);
-
-    let accessToken = params.access_token || params.accessToken;
-
-    const urlParams = new URLSearchParams(hash);
-    console.log("URLSearchParams access_token:", urlParams.get("access_token"));
-    console.log("URLSearchParams accessToken:", urlParams.get("accessToken"));
-    console.log("All URLSearchParams:", [...urlParams.entries()]);
-
-    if (!accessToken) {
-      accessToken =
-        urlParams.get("access_token") || urlParams.get("accessToken");
-    }
-
-    console.log("Final accessToken:", accessToken);
-
-    if (!accessToken) {
-      console.log("No access token found with either naming convention");
-      return false;
-    }
-
-    setIsProcessing(true);
-    setError(null);
-
-    try {
-      const suapResponse = await fetch("https://suap.ifpr.edu.br/api/eu/", {
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-          Accept: "application/json",
-        },
-      });
-
-      if (!suapResponse.ok) {
-        throw new Error(`SUAP API error: ${suapResponse.status}`);
-      }
-
-      const suapUserData = await suapResponse.json();
-      const processedUserData = {
-        ra: suapUserData.identificacao,
-        name: suapUserData.nome,
-        email: suapUserData.email,
-        secondaryEmail: suapUserData.email_secundario,
-        campus: suapUserData.campus,
-        userType: suapUserData.tipo_usuario,
-        fullName: suapUserData.nome_registro,
-        firstName: suapUserData.primeiro_nome,
-        lastName: suapUserData.ultimo_nome,
-        photo: suapUserData.foto,
-        role: determineRole(suapUserData),
-      };
-
-      const authResponse = await fetch("/api/auth/suap-login", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          userData: processedUserData,
-          suapToken: accessToken,
-          provider: "SUAP_IFPR",
-        }),
-      });
-
-      if (!authResponse.ok) {
-        throw new Error("Backend authentication failed");
-      }
-
-      const authResult = await authResponse.json();
-
-      localStorage.setItem("divulgaifToken", authResult.token);
-      localStorage.setItem("divulgaifUser", JSON.stringify(authResult.user));
-
-      try {
-        const response = await api.post(BASE_URL + "/api/v1/users", {
-          name: processedUserData.name,
-          email: processedUserData.email,
-          secondaryEmail: processedUserData.secondaryEmail,
-          ra: processedUserData.ra,
-          avatarUrl: processedUserData.photo,
-          userType: processedUserData.userType,
-        });
-
-        console.log("User created:", response.data);
-      } catch (error) {
-        console.error("Failed to create user:", error);
-      }
-
-      window.history.replaceState({}, document.title, window.location.pathname);
-
-      navigate(`/${i18n.language}`);
-
-      return true;
-    } catch (err) {
-      setError("Falha na autenticação com SUAP. Tente novamente.");
-      window.history.replaceState({}, document.title, window.location.pathname);
-
-      return false;
-    } finally {
-      setIsProcessing(false);
-    }
-  };
-
-  const determineRole = (suapData) => {
-    if (suapData.tipo_usuario.includes("aluno")) {
-      return "STUDENT";
-    }
-    return suapData.tipo_usuario === "Aluno" ? "STUDENT" : "TEACHER";
-  };
 
   const handleApplyFilters = (filters) => {
     setActiveFilters(filters);
